04/16/24

+++ include/resource_functions.php
@@ -5016,10 +5016,23 @@ function add_alternative_file($resource,$name,$description="",$file_name="",$fil
     return sql_insert_id();
     }
 
+/**
+ * delete_alternative_file
+ *
+ * @param   int   $resource   Resource id.
+ * @param   int   $ref        Alternative file id.
+ * 
+ * @return  bool   false on error else true.
+ */
 function delete_alternative_file($resource,$ref)
     {
     # Delete any uploaded file.
     $info=get_alternative_file($resource,$ref);
+    if (!$info)
+        {
+        return false;
+        }
+
     $path=get_resource_path($resource, true, "", true, $info["file_extension"], -1, 1, false, "", $ref);
     hook('delete_alternative_file_extra', '', array($path));
     if (file_exists($path)) {unlink($path);}
@@ -5206,8 +5219,7 @@ function get_keyword_from_option($option)
 
 function get_resource_access($resource)
     {
-    global $customgroupaccess,$customuseraccess, $internal_share_access, $k,$uploader_view_override, $userref,
-        $prevent_open_access_on_edit_for_active, $open_access_for_contributor,
+    global $customgroupaccess,$customuseraccess, $internal_share_access, $k,$uploader_view_override, $userref, $open_access_for_contributor,
         $userref,$usergroup, $usersearchfilter, $search_all_workflow_states,
         $userderestrictfilter, $userdata, $custom_access;
     # $resource may be a resource_data array from a search, in which case, many of the permissions checks are already done.
@@ -5316,7 +5328,7 @@ function get_resource_access($resource)
             }
         }
 
-    if ($access == 1 && get_edit_access($ref,$resourcedata['archive'],$resourcedata) && !$prevent_open_access_on_edit_for_active)
+    if ($access == 1 && get_edit_access($ref,$resourcedata['archive'],$resourcedata))
         {
         # If access is restricted and user has edit access, grant open access.
         $access = 0;
